+++ inc/utils.hpp
@@ -13,5 +13,5 @@ bool isCmd(std::string cmd);
 bool isMode(char m);
 
 int toInt(std::string& input);
-std::string toString(size_t& input);
+std::string toString(int& input);
